feat: 리워드 메커니즘 변경 — upload 0.5pt, comment 0.1pt (v0.32.0)
- POINTS_PER_UPLOAD 50 → 0.5, POINTS_PER_COMMENT 0.1 신규
- like_received / view_received 리워드 제거
- reward_points.points 컬럼 Integer → Float (마이그레이션 포함)
- 댓글 작성 시 commenter에게 0.1pt 즉시 지급
- 프론트엔드: 대기 포인트 표시 ÷100 제거, L → pt

+++ backend/app/models/reward.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String, func
+from sqlalchemy import DateTime, Float, ForeignKey, Integer, String, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -14,8 +14,8 @@ class RewardPoint(Base):
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
     user_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.id"), nullable=False)
     week_label: Mapped[str] = mapped_column(String, nullable=False)  # e.g. "2026-W21"
-    points: Mapped[int] = mapped_column(Integer, nullable=False)
-    reason: Mapped[str] = mapped_column(String, nullable=False)  # upload | like_received | view_received
+    points: Mapped[float] = mapped_column(Float, nullable=False)
+    reason: Mapped[str] = mapped_column(String, nullable=False)  # upload | comment
     reference_id: Mapped[int | None] = mapped_column(Integer, nullable=True)
     status: Mapped[str] = mapped_column(String, default="fixed", nullable=False)  # queued | fixed | revoked
     created_at: Mapped[datetime] = mapped_column(
